requirements.txt
kafka-python==2.0.2
textblob==0.15.3
pandas==0.25.3
